+++ CompZoneAerienne/src/CVecZoneReduce.h
@@ -4,7 +4,7 @@
 /// \brief
 ///
 /// \date creation     : 25/11/2024
-/// \date modification : 29/11/2024
+/// \date modification : 30/11/2024
 ///
 
 #ifndef _VECZONEREDUCE_
@@ -29,8 +29,8 @@ public :
     /// \brief un point du polygone
     struct st_coord_poly
         {
-        float m_Lat ;   ///< latitude en degres
-        float m_Lon ;   ///< longitude en degres
+        double m_Lat ;   ///< latitude en degres
+        double m_Lon ;   ///< longitude en degres
         bool  m_PtADetuire = false ;
         bool  m_PtAppartientAUnGroup = false ;
         } ;
@@ -38,8 +38,8 @@ public :
     void Set( std::vector<st_coord_poly*> & VecToReduce )
         { m_pVecOrigine = & VecToReduce ; } ;
     void ReduceToDistanceDroiteAngleDistancePoint( double DistanceMetresEcartDroite , double MemeDirectionEnDegres , double DistanceMetresEntrePoints ) ;
-    int  ReduceMultiDistanceDroite( float DistanceMetresEcartDroite ) ;
-    void ReduceNuageDroite( float DistanceMetresEcartDroite ) ;
+    int  ReduceMultiDistanceDroite( double DistanceMetresEcartDroite ) ;
+    void ReduceNuageDroite( double DistanceMetresEcartDroite ) ;
     void ReduceNuageBravaisPearson( double DistanceMetresEcartDroite , double CoefBravaisPearson ) ;
 
 private :
@@ -47,12 +47,12 @@ private :
     void AjoutePointSansDoublons( std::vector<st_coord_poly*> & VecReduce , int ipa ) ;
     void AjoutePointSansDoublons( std::vector<st_coord_poly*> & VecReduce , st_coord_poly* pStCoord ) ;
 
-    //float GetBravaisPearson( const std::vector<long> &VecIPts ) ;
+    //double GetBravaisPearson( const std::vector<long> &VecIPts ) ;
 
     /*struct st_pts_i_b
         {
         int   m_ipts ;          ///< position vecteur origine
-        float m_DistanceBari ;  ///< distance baricentre
+        double m_DistanceBari ;  ///< distance baricentre
 
         bool operator > ( const st_pts_i_b & pts ) const
             { return m_DistanceBari > pts.m_DistanceBari ; } ;
