+++ CompZoneAerienne/src/CompZoneAerienne.h
@@ -4,7 +4,7 @@
 /// \brief
 ///
 /// \date 26/11/2024 : creation
-/// \date 29/11/2024 : modification
+/// \date 30/11/2024 : modification
 ///
 
 #ifndef _BVTZonesAerienne_
@@ -19,20 +19,21 @@
 #include <algorithm>
 #include <limits>
 #include <cfloat>
+#include <unistd.h>
 
 ///////////
 // macro //
 ///////////
 
-#define FABS(x)     fabsf(x)
-#define SQRT(x)     sqrtf(x)
-#define SIN(x)      sinf(x)
-#define COS(x)      cosf(x)
-#define ACOS(x)     acosf(x)
-#define POW(x,y)    powf(x,y)
-#define FMOD(x,y)   fmodf(x,y)
-#define FLOOR(x)    floorf(x)
-#define MODF(x,y)   modff(x,y)
+#define FABS(x)     fabs(x)
+#define SQRT(x)     sqrt(x)
+#define SIN(x)      sin(x)
+#define COS(x)      cos(x)
+#define ACOS(x)     acos(x)
+#define POW(x,y)    pow(x,y)
+#define FMOD(x,y)   fmod(x,y)
+#define FLOOR(x)    floor(x)
+#define MODF(x,y)   modf(x,y)
 
 #define MIN( a , b )   (((a)<(b)) ? (a) : (b))
 #define MAX( a , b )   (((a)>(b)) ? (a) : (b))
@@ -52,6 +53,7 @@
 #include "CVecZoneReduce.h"
 #include "CNuage2Droite.h"
 #include "CGroupeAligne.h"
+#include "CCompZoneErr.h"
 
 #define IN_ZONE_FILE    "data/20240615_ffvl-cfd.geojson"
 #define UnMilesEnMetres (1852.)
